chore: move PSBTBuilder to new directory
