+++ packages/sdk/src/index.ts
@@ -36,9 +36,9 @@ export * from "./inscription"
 export * from "./instant-trade"
 export * from "./keys"
 export * from "./modules"
+export * from "./psbt-builder"
 export * from "./signatures"
 export * from "./transactions"
-export { PSBTBuilder } from "./transactions/PSBTBuilder"
 export * from "./utils"
 export { UTXOManager } from "./utxos"
 export * from "./wallet"
